Plot covid-19 time series

Takes the data from updateCovid function and passes it on to plot a time series plot for every type of data available.

covidTimeSeries(
  country = "egypt",
  min.date = "2020-01-22",
  max.date = as.Date(today())
)

Arguments

country
name of the country

min.date
a value from 2020-1-22 to today's date

max.date
a value from 2020-1-22 to today's date

Value

a time series plot

Author

Ali Malek

Examples

covidTimeSeries(country = "egypt",min.date = "2020-02-22",max.date = as.Date(today()))

#> Loading required package: tidyverse

#> -- Attaching packages -------------------------------------------------------------------- tidyverse 1.3.0 --

#> v ggplot2 3.3.2     v purrr   0.3.4
#> v tibble  3.0.3     v dplyr   1.0.2
#> v tidyr   1.1.1     v stringr 1.4.0
#> v readr   1.3.1     v forcats 0.5.0

#> -- Conflicts ----------------------------------------------------------------------- tidyverse_conflicts() --
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

#> Error in today(): could not find function "today"